Avoid reflection; use interfaces
I think reflection was used here because 'tool' wasn't available until Init had been run. Instead, the `NewClient` invocation is hoisted to the global variable definition, and the reflection becomes unnecessary.

The motivation for this is to complete the rename of 'test' to 'simulation': gorename can't see through the reflection.